Moved factory methods to helper classes
